2009-11-16  Jeffrey Stedfast  <fejj@novell.com>
	* mp3.cpp (mpeg_frame_length): Use integer math to calculate the
	frame length.
	(ReadHeader): Use integer math to estimate the number of frames in
	the mpeg stream.


svn path=/trunk/moon/; revision=146274

+++ src/mp3.h
@@ -132,6 +132,6 @@ public:
 };
 
 bool mpeg_parse_header (MpegFrameHeader *mpeg, const guint8 *buffer);
-double mpeg_frame_length (MpegFrameHeader *mpeg, bool xing);
+guint32 mpeg_frame_length (MpegFrameHeader *mpeg, bool xing);
 
 #endif
